Extracted tab panel component into its own separate file.
